# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
      "title": "Symfony API REST",
      "description": "CRUD API in Symfony",
      "version": "1.0.0"
  },
  "paths": {
      "/api/categories/get": {
          "get": {
              "operationId": "get_api_categoriesget_all",
              "responses": {
                  "default": {
                      "description": ""
                  }
              },
              "security": [
                {
                    "bearerAuth": []
                }
              ]
          }
      },
      "/api/categories/{category}/get": {
          "get": {
              "operationId": "get_api_categoriesget_by_id",
              "parameters": [
                  {
                      "name": "category",
                      "in": "path",
                      "required": true,
                      "schema": {
                          "type": "string"
                      }
                  }
              ],
              "responses": {
                  "default": {
                      "description": ""
                  }
              }
          }
      },
      "/api/categories/create": {
          "post": {
              "operationId": "post_api_categoriescreate_category",
              "requestBody": {
                  "required": true,
                  "content": {
                      "application/json": {
                          "schema": {
                              "$ref": "#/components/schemas/CategoryDto"
                          }
                      }
                  }
              },
              "responses": {
                  "default": {
                      "description": ""
                  }
              }
          }
      },
      "/api/categories/{id}/update": {
          "put": {
              "operationId": "put_api_categoriesupdate_category",
              "parameters": [
                  {
                      "name": "id",
                      "in": "path",
                      "required": true,
                      "schema": {
                          "type": "string"
                      }
                  }
              ],
              "requestBody": {
                  "required": true,
                  "content": {
                      "application/json": {
                          "schema": {
                              "$ref": "#/components/schemas/CategoryDto"
                          }
                      }
                  }
              },
              "responses": {
                  "default": {
                      "description": ""
                  }
              }
          }
      },
      "/api/categories/{category}/delete": {
          "delete": {
              "operationId": "delete_api_categoriesdelete_category",
              "parameters": [
                  {
                      "name": "category",
                      "in": "path",
                      "required": true,
                      "schema": {
                          "type": "string"
                      }
                  }
              ],
              "responses": {
                  "default": {
                      "description": ""
                  }
              }
          }
      },
...
  },
  "components": {
      "schemas": {
          "CategoryDto": {
              "required": [
                  "name"
              ],
              "properties": {
                  "name": {
                      "type": "string"
                  },
                  "description": {
                      "type": "string",
                      "default": "",
                      "nullable": true
                  }
              },
              "type": "object"
          },
...
      }
  }
}